Writing to both database and iRODS config files under the Tomcat directory (using db and irods specs)

+++ packaging/scripts/install_mlx.py
@@ -11,16 +11,25 @@ from utils import *
 class MetalnxContext(DBConnectionTestMixin, IRODSConnectionTestMixin, FileManipulationMixin):
     def __init__(self):
         self.jar_path = '/usr/bin/jar'
+
+        # Tomcat config params
         self.tomcat_home = '/usr/share/tomcat'
+        self.tomcat_bin_dir = ''
+        self.tomcat_webapps_dir = ''
+        self.classes_path = ''
 
-        self.existing_conf = False
-
+        # Metalnx config params
         self.metalnx_war_path = '/tmp/emc-tmp/emc-metalnx-web.war'
-        self.metalnx_db_properties_path = 'database.properties'
-        self.metalnx_irods_properties_path = 'irods.environment.properties'
+        self.metalnx_db_properties_path = ''
+        self.metalnx_irods_properties_path = ''
+        self.tmp_dir = None
 
+        # Metalnx properties
         self.irods_props = {}
         self.db_props = {}
+
+        # Existing Metalnx configuration flag
+        self.existing_conf = False
 
     def config_java_devel(self):
         """It will make sure the java-devel package is correctly installed"""
@@ -38,18 +47,22 @@ class MetalnxContext(DBConnectionTestMixin, IRODSConnectionTestMixin, FileManipu
         self.tomcat_webapps_dir = path.join(self.tomcat_home, 'webapps')
 
         # If all paths are valid, then this is a valid tomcat directory
-        if self._is_dir_valid(self.tomcat_home) and self._is_dir_valid(self.tomcat_bin_dir) \
-                and self._is_dir_valid(self.tomcat_webapps_dir):
-            return True
+        if not self._is_dir_valid(self.tomcat_home) or not self._is_dir_valid(self.tomcat_bin_dir) \
+                or not self._is_dir_valid(self.tomcat_webapps_dir):
+            raise Exception('Tomcat directory is not valid. Please check the path and try again.')
 
-        raise Exception('Tomcat directory is not valid. Please check the path and try again.')
+        return True
 
     def config_metalnx_package(self):
         """It will check if the Metalnx package has been correctly installed"""
-        if self._is_file_valid(self.metalnx_war_path):
-            return True
 
-        raise Exception('Could not find Metalnx WAR file. Check if emc-metalnx-web package is installed.')
+        print self.metalnx_war_path
+        print self._is_file_valid(self.metalnx_war_path)
+
+        if not self._is_file_valid(self.metalnx_war_path):
+            raise Exception('Could not find Metalnx WAR file. Check if emc-metalnx-web package is installed.')
+
+        return True
 
     def config_existing_setup(self):
         """It will save your current installed of metalnx and will restore them after update"""
@@ -73,8 +86,9 @@ class MetalnxContext(DBConnectionTestMixin, IRODSConnectionTestMixin, FileManipu
         """The installation process will now handle your new WAR file"""
         metalnx_web_dir = path.join(self.tomcat_webapps_dir, 'emc-metalnx-web')
 
-        log('Removing current Metalnx installation directory')
-        rmtree(metalnx_web_dir)
+        if self.existing_conf:
+            log('Removing current Metalnx installation directory')
+            rmtree(metalnx_web_dir)
 
         log('Creating new Metalnx directory')
         mkdir(metalnx_web_dir)
@@ -112,10 +126,12 @@ class MetalnxContext(DBConnectionTestMixin, IRODSConnectionTestMixin, FileManipu
 
                 self.irods_props[spec['name']] = response
 
-            print self.irods_props
-
         self._test_irods_connection()
-        # self._write_irods_properties_to_file()
+
+        self.irods_props[IRODS_PROPS_SPEC['password']['name']] = IRODS_PROPS_SPEC['password']['cast'](
+            self.irods_props[IRODS_PROPS_SPEC['password']['name']])
+
+        self._write_irods_properties_to_file()
 
     def config_database(self):
         """It will configure database access"""
@@ -133,34 +149,43 @@ class MetalnxContext(DBConnectionTestMixin, IRODSConnectionTestMixin, FileManipu
 
                 self.db_props[spec['name']] = response
 
-            db_url = DB_TYPE_SPEC['url']['cast'](
-                db_type,
-                self.db_props[DB_PROPS_SPEC['host']['name']],
-                self.db_props[DB_PROPS_SPEC['port']['name']],
-                self.db_props[DB_PROPS_SPEC['db_name']['name']]
-            )
+        # Database URL connection
+        db_url = DB_TYPE_SPEC['url']['cast'](
+            db_type,
+            self.db_props[DB_PROPS_SPEC['host']['name']],
+            self.db_props[DB_PROPS_SPEC['port']['name']],
+            self.db_props[DB_PROPS_SPEC['db_name']['name']]
+        )
 
-            self.db_props[DB_TYPE_SPEC['url']['name']] = db_url
-            self.db_props.update(DB_TYPE_SPEC[db_type])
+        self.db_props[DB_TYPE_SPEC['url']['name']] = db_url
+        self.db_props.update(DB_TYPE_SPEC[db_type])
 
-            self._test_database_connection(db_type)
+        # Hibernate config params
+        self.db_props.update(HIBERNATE_CONFIG[db_type])
+        self.db_props.update(HIBERNATE_CONFIG['options'])
 
-            del self.db_props[DB_PROPS_SPEC['host']['name']]
-            del self.db_props[DB_PROPS_SPEC['port']['name']]
-            del self.db_props[DB_PROPS_SPEC['db_name']['name']]
+        self._test_database_connection(db_type)
 
-            # self._write_db_properties_to_file()
+        # props used to build the DB url connection but do not exist in the database properties file
+        del self.db_props[DB_PROPS_SPEC['host']['name']]
+        del self.db_props[DB_PROPS_SPEC['port']['name']]
+        del self.db_props[DB_PROPS_SPEC['db_name']['name']]
+
+        self.db_props[DB_PROPS_SPEC['password']['name']] = DB_PROPS_SPEC['password']['cast'](
+            self.db_props[DB_PROPS_SPEC['password']['name']])
+
+        self._write_db_properties_to_file()
 
     def config_restore_conf(self):
         """Restoring existing Metalnx configuration"""
         if self.existing_conf:
             self._move_properties_files(self.tmp_dir, self.classes_path)
 
-            # Removign temp directory
+            # Removing temp directory
             rmtree(self.tmp_dir)
 
     def config_set_https(self):
-        """Sets HTTPS protocol in Tomcat and Metanlx"""
+        """Sets HTTPS protocol in Tomcat and Metalnx"""
         metalnx_web_xml = path.join(self.tomcat_webapps_dir, 'emc-metalnx-web', 'WEB-INF', 'classes', 'web.xml')
         tomcat_server_xml = path.join(self.tomcat_home, 'conf', 'server.xml')
